Final reorg with executable and README
