+++ README.md
@@ -1,2 +1,32 @@
 # tic_tac_toe
-Code challenge for BYU OIT application
+Code challenge for BYU OIT application.
+
+## Notes
+
+I solved this challenge in Typescript, which is a superset of Javascript. I chose to use Typescript because it is a statically typed language, which means that it can catch errors at compile time rather than at runtime. This makes it easier to catch bugs and write more reliable code. I also chose to use Typescript because it is a language that I am familiar with and enjoy using.
+
+The game is played on the command line. Since there isn't a 'clicking' option the user inputs a number from 1-9 to place their piece on the board. The board is displayed as follows:
+
+```
+1 | 2 | 3
+---------
+4 | 5 | 6
+---------
+7 | 8 | 9
+```
+
+## Design
+
+I used a pretty bare-bones design for my solution. Since it was a pretty simple game loop and didn't require a lot of stored data between turns, I didn't see a need for a lot of different classes or functions. I used a pretty linear approach to the game loop and the functions required. I also used some helper functions for input validation, board checks, user prompts, and displaying the board.
+
+## External Dependencies
+
+I didn't need any external dependencies besdies those required for TypeScript. I used the `readline` module from Node.js to get user input from the command line.
+
+## Challenge Details
+
+I spent about 3.5  hours on this challenge. I spent about 1 hour planning and designing the solution, about 2 hours coding the solution, and about 30 minutes testing and debugging the solution. I didn't run into any major issues while working on the challenge. I was able to complete the challenge in the time that I had set aside for it.
+
+## Thank You
+
+Thanks for the opportunity to apply to OIT again! I hope things turn out well and I am able to join the team. I look forward to hearing back from you soon. Have a great day!
